Rename TracerFactory to TracerProvider
This follows opentelemetry/oteps#76.

Also start moving sdk tests into the sdk/Tests folder
