+++ api/Trace/TracerProvider.php
@@ -4,7 +4,7 @@ declare(strict_types=1);
 
 namespace OpenTelemetry\Trace;
 
-interface TracerFactory
+interface TracerProvider
 {
     public function getTracer(string $name, ?string $version = null): Tracer;
 }
