+++ sdk/Trace/TracerProvider.php
@@ -6,7 +6,7 @@ namespace OpenTelemetry\Sdk\Trace;
 
 use OpenTelemetry\Trace as API;
 
-class TracerFactory
+class TracerProvider implements API\TracerProvider
 {
     /**
      * @var self
@@ -31,7 +31,7 @@ class TracerFactory
     final private function __construct(array $spanProcessors = [])
     {
         foreach ($spanProcessors as $spanProcessor) {
-            if (!$spanProcessor instanceof API\SpanProcessor) {
+            if (!$spanProcessor instanceof SpanProcessor) {
                 throw new \TypeError(
                     sprintf(
                         'Span Processors should be of type %s, but object of type %s provided',
@@ -52,16 +52,16 @@ class TracerFactory
      */
     public static function getInstance(array $spanProcessors = []): self
     {
-        if (self::$instance instanceof TracerFactory) {
+        if (self::$instance instanceof TracerProvider) {
             return self::$instance;
         }
 
-        $instance = new TracerFactory($spanProcessors);
+        $instance = new TracerProvider($spanProcessors);
 
         return self::$instance = $instance;
     }
 
-    public function getTracer(string $name, string $version = ''): API\Tracer
+    public function getTracer(string $name, ?string $version = ''): API\Tracer
     {
         if (isset($this->tracers[$name]) && $this->tracers[$name] instanceof API\Tracer) {
             return $this->tracers[$name];
